create bill

+++ SourceDoAn/SML_QLNPP/DataService/Interfaces/IDeliveryOrderService.cs
@@ -11,7 +11,7 @@ namespace DataService.Interfaces
         bool AddDeliveryOrder(DeliveryOrder dOrder);
         bool UpdateDeliveryOrder(DeliveryOrder dOrder);
         DeliveryOrder SearchById(int id);
-        IList<DeliveryOrder> SearchByStatus(int status);
+        IList<DeliveryOrder> SearchByStatus(byte status);
         IList<DeliveryOrder> SearchByDeliveryDate(DateTime datetime);
 
         // can them vao
